+++ Objective-C/TOCropViewController/TOCropViewController.m
@@ -678,10 +678,15 @@ static const CGFloat kTOCropViewControllerToolbarHeight = 44.0f;
             break;
     }
     
+    // If the aspect ratio lock is not enabled, allow a swap
+    // If the aspect ratio lock is on, allow a aspect ratio swap
+    // only if the allowDimensionSwap option is specified.
+    BOOL aspectRatioCanSwapDimensions = !self.aspectRatioLockEnabled ||
+                                (self.aspectRatioLockEnabled && self.aspectRatioLockDimensionSwapEnabled);
+    
     //If the image is a portrait shape, flip the aspect ratio to match
-    if (aspectRatioPreset != TOCropViewControllerAspectRatioPresetCustom &&
-        self.cropView.cropBoxAspectRatioIsPortrait &&
-        !self.aspectRatioLockEnabled)
+    if (self.cropView.cropBoxAspectRatioIsPortrait &&
+        aspectRatioCanSwapDimensions)
     {
         CGFloat width = aspectRatio.width;
         aspectRatio.width = aspectRatio.height;